Use the correct target for INVOKESPECIAL when calling default methods The target of the invocation is the closest interface in the hierarchy no matter in which interface the default method is defined.
Example: java.util.Collection defines a stream() default method
If you create a class implementing the List interface, which is a
subtype of Collection, the correct target for INVOKESPECIAL is
List.stream(), not Collection.stream().

Note that this implementation relies on the fact that Mockito's
ClassImposterizer puts the mocked interface at the first position in the
interface list of the ClassInfo.